+++ internal/domain/handlers/cart_handler.go
@@ -11,7 +11,7 @@ type Cart = service.Cart // declare type models Cart
 HANDLER GET CART
 */
 func GetCart(c *fiber.Ctx) error {
-	id := c.Locals("user_id").(uint) // get user id
+	id := c.Locals("user_id").(uint64) // get user id
 	cart := service.GetCart(id)
 	return c.Status(200).JSON(cart)
 }
@@ -21,11 +21,11 @@ HANDLER FIND CART
 */
 func FindCart(c *fiber.Ctx) error {
 	// get user id & cart id
-	user_id := c.Locals("user_id").(uint)
+	user_id := c.Locals("user_id").(uint64)
 	id, _ := c.ParamsInt("id")
 
 	// service Find a Cart
-	cart, _ := service.FindCart(uint(id))
+	cart, _ := service.FindCart(uint64(id))
 
 	// check cart exist
 	if cart.ID == 0 {
@@ -48,7 +48,7 @@ HANDLER CREATE CART
 */
 func CreateCart(c *fiber.Ctx) error {
 	var cart Cart                         // declare variabel Cart
-	id_user := c.Locals("user_id").(uint) // get user id
+	id_user := c.Locals("user_id").(uint64) // get user id
 
 	// Bind body request ke struct cart
 	if err := c.BodyParser(&cart); err != nil {
@@ -64,14 +64,14 @@ func CreateCart(c *fiber.Ctx) error {
 	}
 
 	// check product
-	if uint(product.Stock) < 1 {
+	if uint64(product.Stock) < 1 {
 		return c.Status(400).JSON(fiber.Map{
 			"message": "Stock Product is empty",
 		})
 	}
 
 	// data total_cost
-	var cost uint = product.Price
+	var cost uint64 = product.Price
 
 	// Connect service
 	carts := service.CreateCart(cart, id_user, cost)
@@ -83,7 +83,7 @@ HANDLER UPDATE CART
 */
 func UpdateCart(c *fiber.Ctx) error {
 	var cart Cart                         // declare variabel Cart
-	user_id := c.Locals("user_id").(uint) // get user id
+	user_id := c.Locals("user_id").(uint64) // get user id
 
 	// Bind body request ke struct cart
 	if err := c.BodyParser(&cart); err != nil {
@@ -135,17 +135,17 @@ HANDLER DELETE CART
 func DeleteCart(c *fiber.Ctx) error {
 	// Get id endpoint & user id
 	id, _ := c.ParamsInt("id")
-	user_id := c.Locals("user_id").(uint)
+	user_id := c.Locals("user_id").(uint64)
 
 	// Find Cart
-	carts, _ := service.FindCart(uint(id))
+	carts, _ := service.FindCart(uint64(id))
 	if carts.UserID != user_id {
 		return c.Status(403).JSON(fiber.Map{
 			"message": "Forbidden",
 		})
 	}
 	// Delete cart
-	service.DeleteCart(uint(id))
+	service.DeleteCart(uint64(id))
 	return c.Status(200).JSON(fiber.Map{
 		"message": "Success",
 	})
